refactor: Extract documentation generator + fix include/exclude
