+++ package.json
@@ -27,8 +27,6 @@
   "dependencies": {
     "debug": "^4.1.1",
     "endent": "^2.0.1",
-    "find-cache-dir": "^3.3.1",
-    "flat-cache": "^3.0.4",
     "micromatch": "^4.0.2",
     "react-docgen-typescript": "^1.20.5",
     "tslib": "^2.0.0",
