+++ src/plugin.ts
@@ -6,20 +6,11 @@ import * as webpack from "webpack";
 import ts from "typescript";
 import * as docGen from "react-docgen-typescript";
 import { matcher } from "micromatch";
-import findCacheDir from "find-cache-dir";
-import flatCache from "flat-cache";
-import crypto from "crypto";
 
+import { LoaderOptions } from "./types";
 import DocGenDependency from "./dependency";
-import { generateDocgenCodeBlock } from "./generateDocgenCodeBlock";
 
 const debugExclude = createDebug("docgen:exclude");
-const debugInclude = createDebug("docgen:include");
-const debugDocs = createDebug("docgen:docs");
-
-const cacheId = "ts-docgen";
-const cacheDir = findCacheDir({ name: cacheId });
-const cache = flatCache.load(cacheId, cacheDir);
 
 interface TypescriptOptions {
   /**
@@ -31,41 +22,6 @@ interface TypescriptOptions {
   compilerOptions?: ts.CompilerOptions;
 }
 
-interface LoaderOptions {
-  /**
-   * Specify the docgen collection name to use. All docgen information will
-   * be collected into this global object. Set to null to disable.
-   *
-   * @default STORYBOOK_REACT_CLASSES
-   * @see https://github.com/gongreg/react-storybook-addon-docgen
-   **/
-  docgenCollectionName?: string | null;
-
-  /**
-   * Automatically set the component's display name. If you want to set display
-   * names yourself or are using another plugin to do this, you should disable
-   * this option.
-   *
-   * ```
-   * class MyComponent extends React.Component {
-   * ...
-   * }
-   *
-   * MyComponent.displayName = "MyComponent";
-   * ```
-   *
-   * @default true
-   */
-  setDisplayName?: boolean;
-
-  /**
-   * Specify the name of the property for docgen info prop type.
-   *
-   * @default "type"
-   */
-  typePropName?: string;
-}
-
 export type PluginOptions = docGen.ParserOptions &
   LoaderOptions &
   TypescriptOptions & {
@@ -74,65 +30,6 @@ export type PluginOptions = docGen.ParserOptions &
     /** Glob patterns to include. defaults to ts|tsx */
     include?: string[];
   };
-
-interface Module {
-  userRequest: string;
-  request: string;
-  built?: boolean;
-  rawRequest?: string;
-  external?: boolean;
-  _source: {
-    _value: string;
-  };
-}
-
-/** Run the docgen parser and inject the result into the output */
-function processModule(
-  parser: docGen.FileParser,
-  webpackModule: Module,
-  tsProgram: ts.Program,
-  loaderOptions: Required<LoaderOptions>
-) {
-  if (!webpackModule) {
-    return;
-  }
-
-  const hash = crypto
-    .createHash("sha1")
-    .update(webpackModule._source._value)
-    .digest("hex");
-  const cached = cache.getKey(hash);
-
-  if (cached) {
-    debugInclude(`Got cached docgen for "${webpackModule.request}"`);
-    webpackModule._source._value = cached;
-    return;
-  }
-
-  const componentDocs = parser.parseWithProgramProvider(
-    webpackModule.userRequest,
-    () => tsProgram
-  );
-
-  if (!componentDocs.length) {
-    return;
-  }
-
-  const docs = generateDocgenCodeBlock({
-    filename: webpackModule.userRequest,
-    source: webpackModule.userRequest,
-    componentDocs,
-    ...loaderOptions,
-  }).substring(webpackModule.userRequest.length);
-
-  debugDocs(docs);
-
-  let sourceWithDocs = webpackModule._source._value;
-  sourceWithDocs += `\n${docs}\n`;
-  webpackModule._source._value = sourceWithDocs;
-
-  cache.setKey(hash, sourceWithDocs);
-}
 
 /** Get the contents of the tsconfig in the system */
 function getTSConfigFile(tsconfigPath: string): ts.ParsedCommandLine {
@@ -153,8 +50,8 @@ function getTSConfigFile(tsconfigPath: string): ts.ParsedCommandLine {
 }
 
 /** Create a glob matching function. */
-const matchGlob = (globs: string[]) => {
-  const matchers = globs.map((g) => matcher(g));
+const matchGlob = (globs?: string[]) => {
+  const matchers = (globs || []).map((g) => matcher(g));
   return (filename: string) =>
     Boolean(filename && matchers.find((match) => match(filename)));
 };
@@ -197,6 +94,10 @@ export default class DocgenPlugin {
   apply(compiler: webpack.Compiler): void {
     const pluginName = "DocGenPlugin";
 
+    const { exclude, include } = this.options;
+    const isExcluded = matchGlob(exclude);
+    const isIncluded = matchGlob(include);
+
     compiler.hooks.compilation.tap(
       pluginName,
       (compilation, { normalModuleFactory }) => {
@@ -204,7 +105,7 @@ export default class DocgenPlugin {
           // eslint-disable-next-line
           // @ts-ignore TODO: Figure out why this isn't allowed
           DocGenDependency,
-          new DocGenDependency.Template()
+          new DocGenDependency.Template(this.parser)
         );
 
         // eslint-disable-next-line
@@ -234,22 +135,19 @@ export default class DocgenPlugin {
               return;
             }
 
-            // TODO: Re-enable these two
-            /*
             if (isExcluded(module.userRequest)) {
               debugExclude(
                 `Module not matched in "exclude": ${module.userRequest}`
               );
               return;
             }
-  
+
             if (!isIncluded(module.userRequest)) {
               debugExclude(
                 `Module not matched in "include": ${module.userRequest}`
               );
               return;
             }
-            */
 
             // eslint-disable-next-line
             // @ts-ignore
@@ -268,79 +166,6 @@ export default class DocgenPlugin {
           .tap(pluginName, handler);
       }
     );
-  }
-
-  // TODO: Eliminate this one after the new apply works
-  oldApply(compiler: webpack.Compiler): void {
-    const {
-      docgenCollectionName = "STORYBOOK_REACT_CLASSES",
-      setDisplayName = true,
-      typePropName = "type",
-      exclude = [],
-      include = ["**/**.tsx"],
-    } = this.options;
-
-    const isExcluded = matchGlob(exclude);
-    const isIncluded = matchGlob(include);
-
-    compiler.hooks.make.tap(this.name, (compilation) => {
-      compilation.hooks.seal.tap(this.name, () => {
-        const modulesToProcess: Module[] = [];
-
-        // eslint-disable-next-line
-        // @ts-ignore TODO: Module type
-        compilation.modules.forEach((module: Module) => {
-          if (!module.built) {
-            debugExclude(`Ignoring un-built module: ${module.userRequest}`);
-            return;
-          }
-
-          if (module.external) {
-            debugExclude(`Ignoring external module: ${module.userRequest}`);
-            return;
-          }
-
-          if (!module.rawRequest) {
-            debugExclude(
-              `Ignoring module without "rawRequest": ${module.userRequest}`
-            );
-            return;
-          }
-
-          if (isExcluded(module.userRequest)) {
-            debugExclude(
-              `Module not matched in "exclude": ${module.userRequest}`
-            );
-            return;
-          }
-
-          if (!isIncluded(module.userRequest)) {
-            debugExclude(
-              `Module not matched in "include": ${module.userRequest}`
-            );
-            return;
-          }
-
-          debugInclude(module.userRequest);
-          modulesToProcess.push(module);
-        });
-
-        const tsProgram = ts.createProgram(
-          modulesToProcess.map((v) => v.userRequest),
-          this.compilerOptions
-        );
-
-        modulesToProcess.forEach((m) =>
-          processModule(this.parser, m, tsProgram, {
-            docgenCollectionName,
-            setDisplayName,
-            typePropName,
-          })
-        );
-
-        cache.save();
-      });
-    });
   }
 }
